refactor: Reorganized helpers for test fixtures
