+++ tests/helpers/testing.js
@@ -1,23 +1,9 @@
 import { strict as assert } from 'node:assert';
-import { existsSync, rmSync } from 'node:fs';
 
 import { test } from '@sondr3/minitest';
-import fixturify from 'fixturify';
-
-export function assertFixture(outputProject, options) {
-  const { projectRoot } = options;
-
-  assert.deepStrictEqual(fixturify.readSync(projectRoot), outputProject);
-}
-
-export function loadFixture(inputProject, options) {
-  const { projectRoot } = options;
-
-  if (existsSync(projectRoot)) {
-    rmSync(projectRoot, { recursive: true });
-  }
-
-  fixturify.writeSync(projectRoot, inputProject);
-}
 
 export { assert, test };
+
+export * from './testing/assert-fixture.js';
+export * from './testing/convert-fixture-to-json.js';
+export * from './testing/load-fixture.js';
